software/requirements.txt
numpy
vedo
pyserial
